Removed FBX SDK from main branch
- The FBX SDK was estimated to have a size of 1.08 GB and that doesn't
work for GitHub. Since it's required, I will keep it at all times
available on my USB unit and give everyone a copy of it. The
libfbxsdk.dll must also be placed in the same folder as the solution. I
will look further into this matter. Thanks Jonathan for fixing this!
:+1:

- The previous flickering in the rendering of meshes was caused by
missing constant buffer data, such as the camera position. Once
corrected, the problems disappeared.

- This solution should be cleaned in both Release and Debug mode