Added ignore for FBX SDK Folder and other troublesome files
Hi everyone, I've now received help from teachers and managed to find a
great template for a gitignore file that I've modified to not include
the FBX SDK file when we commit to master.

You must download the new master in order to receive the gitignore file
and once you've done that, we should never have to move the FBX SDK
folder ever again. We also don't need to clean the solution anymore,
since those files are also added in the gitignore file.